feat: elaspedTime 컴포넌트 분리로 렌더링 최적화
